+++ core/src/main/java/arez/ComputedValue.java
@@ -95,7 +95,8 @@ public final class ComputedValue<T>
                                                           o ),
                               priority,
                               false,
-                              observeLowerPriorityDependencies );
+                              observeLowerPriorityDependencies,
+                              false );
     if ( null != _component )
     {
       _component.addComputedValue( this );